Move query parameter decoders up one level
They are not used by the endpoint implementation, but only in the
class that defines the route endpoints for http4s.
